Address PR 456 ReviewGPT migration findings
