+++ agent-docs/references/testing-ci-map.md
@@ -59,7 +59,7 @@ Last verified: 2026-07-06
 - `.github/workflows/cloudflare-hosted-e2e.yml` runs focused hosted local E2E jobs on GitHub-hosted `ubuntu-24.04` for every pull request and `main` push. A shared `runner-bundle` job builds the hosted-local runner bundle plus workspace `dist` outputs once per run with `MURPH_RUNNER_BUNDLE_BUILD_CONCURRENCY=4` and publishes them as a tar.zst artifact; each scenario job downloads that artifact, unpacks it at the repo root, installs the pinned Codex CLI version from `Dockerfile.cloudflare-hosted-runner-base` so the host-side harness can read the bundled model catalog, and runs its scenario with `--no-bundle` instead of rebuilding the workspace (the workspace `dist` outputs ride along because the Temporal worker's workflow webpack bundle resolves `@murphai/*` package exports to `dist/`). Each scenario job provisions a local `postgres:17` service container (pulled from the `public.ecr.aws/docker/library` mirror to avoid anonymous Docker Hub rate limits) with an explicit `pg_isready -U postgres -d murph_test` readiness probe, runs `pnpm temporal:cli:setup` before the managed local Temporal harness starts, does not authenticate to GHCR before running PR-controlled code, uses anonymous pulls for the public fingerprinted native runner base image on cold runners, and points `DATABASE_URL` at `127.0.0.1:5432/murph_test`, while still using deterministic CI-only hosted-web placeholders for hosted device routing, contact privacy, mailbox fingerprinting, and Privy values. The workflow then runs `pnpm hosted-local e2e device-connect`, `pnpm hosted-local e2e codex-image-media-delivery`, `pnpm hosted-local e2e linq-delivery`, `pnpm hosted-local e2e linq-webhook` (the audio-media gate that proves the Worker-mediated Workers AI transcription path through the container parser drain, the remote-transcription provider, and the `murph-transcribe.worker` egress handler with a deterministic fake `AI` binding from the hosted-local test entrypoint), `pnpm hosted-local e2e linq-scheduled-reminder`, `pnpm hosted-local e2e idle-checkpoint-deferred-progress`, `pnpm hosted-local e2e direct-r2-presigned-put`, `pnpm hosted-local e2e temporal-orchestration`, and `pnpm hosted-local e2e telegram`, and always uploads per-job logs plus redacted hosted-local `state.json` files instead of broadening to the full serial hosted-local E2E suite or uploading every harness artifact. The local aggregate `pnpm --dir apps/cloudflare test:e2e:local` also runs the Workers-runtime E2E lane through `test:e2e:workers:local`; CI keeps that narrower Workers proof inside `apps/cloudflare verify` / `test:workers` rather than duplicating it in each hosted-local E2E job.
 - `.github/workflows/cloudflare-hosted-device-sync-e2e.yml` runs the Junction wearable direct-resource replay hosted local E2E job on every pull request and `main` push. The single job uses the same loopback `postgres:17` service (same `public.ecr.aws/docker/library` mirror) and explicit readiness probe, pinned host Codex CLI install for the hosted-local model catalog, Temporal CLI setup, anonymous runner-image pulls, parallel runner-bundle build concurrency, and deterministic hosted-web placeholder env pattern as the hosted E2E workflow, but keeps its own in-job runner bundle build because a separate prepare job would serialize a single-job workflow, then runs `pnpm hosted-local e2e device-sync-junction-wearable-direct-resource-replay` and always uploads the focused job log.
 - `.github/workflows/release.yml` uses GitHub-hosted `ubuntu-24.04`, installs once, runs `pnpm release:check` with `MURPH_TEST_LANES_PARALLEL=1`, `MURPH_APP_VERIFY_PARALLEL=1`, and `MURPH_VERIFY_STEP_PARALLEL=1` so the release verification lane uses the parallel package/smoke branches and parallel app substeps without enabling full app/package overlap unless `MURPH_ACCEPTANCE_APP_VERIFY_WITH_COVERAGE=1` is set explicitly, while the same deterministic hosted-web build placeholders keep `apps/web verify` on its truthful build path without injecting production DB or production hosted device secrets, then packs the publishable tarballs once for upload/publication.
-- Vercel deploys of `apps/web` use the checked-in Vercel build command `pnpm release:production:migrate && pnpm build`, so the guarded migration wrapper still runs automatically on main-branch production deploys while preview/non-main builds skip through the wrapper guard. The generic `pnpm --dir apps/web build` script is non-mutating and does not run production migrations. The guarded predeploy migration entrypoint uses `DIRECT_DATABASE_URL` when present, requires it in Vercel production, rejects known pooled Postgres ports such as `6432` and `6543`, blocks future destructive Prisma migration SQL after `20260707170000_drop_stale_linq_recency_columns`, regenerates the hosted web Prisma client after migrations, and runs the hosted Linq DB configured-line sync/readiness check so DB-backed Linq assignment cannot deploy with an empty assignable line pool or stale generated client. Production deploy sync skips provider inventory; provider inventory remains on explicit operator/contact-card paths. Normal production Prisma migrations must remain backward compatible with the currently deployed app because a successful predeploy migration can outlive a later build failure. Destructive hosted web contract cleanup belongs in `apps/web/prisma/contract-migrations` and is applied by `.github/workflows/hosted-web-contract-migrations.yml` after a successful production deployment status for `main`; that workflow checks out the deployed SHA and requires a GitHub Actions `HOSTED_WEB_DIRECT_DATABASE_URL` secret.
+- Vercel deploys of `apps/web` use the checked-in Vercel build command `pnpm release:production:migrate && pnpm build`, so the guarded migration wrapper still runs automatically on main-branch production deploys while preview/non-main builds skip through the wrapper guard. The generic `pnpm --dir apps/web build` script is non-mutating and does not run production migrations. The guarded predeploy migration entrypoint uses `DIRECT_DATABASE_URL` when present, requires it in Vercel production, rejects known pooled Postgres ports such as `6432` and `6543`, blocks future destructive Prisma migration SQL after `20260707170000_drop_stale_linq_recency_columns`, regenerates the hosted web Prisma client after migrations, and runs the hosted Linq DB configured-line sync/readiness check so DB-backed Linq assignment cannot deploy with an empty assignable line pool or stale generated client. Production deploy sync skips provider inventory; provider inventory remains on explicit operator/contact-card paths. Normal production Prisma migrations must remain backward compatible with the currently deployed app because a successful predeploy migration can outlive a later build failure. Destructive hosted web contract cleanup belongs in `apps/web/prisma/contract-migrations` and is applied by `.github/workflows/hosted-web-contract-migrations.yml` after a successful production deployment status; that workflow checks out the deployed SHA, verifies it is reachable from `origin/main`, and requires a GitHub Actions `HOSTED_WEB_DIRECT_DATABASE_URL` secret.
 
 ## Current Gaps
